backend/requirements.txt
hvac==1.2.1
Flask==2.3.3
python-dotenv==1.0.0
Flask-Cors==4.0.0
